test(events): keep contract gate within package baseline
